fix: Submodule structure

+++ ms-clients/application/usecase/src/main/java/ec/com/banking/usecase/DeleteClientUseCase.java
@@ -1,6 +1,7 @@
 package ec.com.banking.usecase;
 
-import ec.com.banking.gateway.ClientRepository;
+import ec.com.banking.core.gateway.ClientRepository;
+import ec.com.banking.core.exception.ClientNotFoundException;
 
 public class DeleteClientUseCase {
     private final ClientRepository clientRepository;
@@ -10,6 +11,9 @@ public class DeleteClientUseCase {
     }
 
     public void execute(Long clientId) {
+        if(clientRepository.findById(clientId) == null){
+            throw new ClientNotFoundException(clientId);
+        }
         clientRepository.delete(clientId);
     }
 }
